working on Exam's Controller. altering some url and endpoints

+++ Backend/Quiz-App/src/main/java/com/quizapp/Quiz/App/Controller/Exams.java
@@ -2,6 +2,7 @@ package com.quizapp.Quiz.App.Controller;
 
 import com.quizapp.Quiz.App.Entity.exams;
 import com.quizapp.Quiz.App.Services.examService;
+import com.quizapp.Quiz.App.Services.getloginService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,43 +17,61 @@ public class Exams {
     @Autowired
     private examService examservice;
 
+    @Autowired
+    private getloginService getloginservice;
+
     //<----------------------------------------------------------------------------------------->
     //Request Handler for creating new Exam. Parameter values : id=> id of the course. examdata=>data of the exams.
     @PostMapping("/create/{id}")
     public ResponseEntity<?> create(@PathVariable int id, @RequestBody exams examdata) {
         Authentication authentication= SecurityContextHolder.getContext().getAuthentication();
         String email=authentication.getName();
-        System.out.println(id);
-        System.out.println(email);
-        System.out.println(authentication);
-        exams currentsave=examservice.createExam(id, email, examdata);
-        if(currentsave!=null){
-            return new ResponseEntity<>(currentsave, HttpStatus.CREATED);
+        if(getloginservice.getAuth(email)){
+            exams currentsave=examservice.createExam(id, email, examdata);
+            if(currentsave!=null){
+                return new ResponseEntity<>(currentsave, HttpStatus.CREATED);
+            }else{
+                return new ResponseEntity<>(HttpStatus.CONFLICT);
+            }
         }else{
-            return new ResponseEntity<>(HttpStatus.CONFLICT);
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
         }
+
 
     }// Method End --> CreateExam
 
     //<---------------------------------------------------------------------------------->
     @GetMapping("/getexams/{id}")
     public ResponseEntity<?> getexams(@PathVariable int id) {
-        exams exam=examservice.getExam(id);
-        if(exam!=null){
-            return new ResponseEntity<>(exam, HttpStatus.OK);
+        Authentication authentication= SecurityContextHolder.getContext().getAuthentication();
+        String email=authentication.getName();
+        if(getloginservice.getAuth(email)){
+            exams exam=examservice.getExam(id, email);
+            if(exam!=null){
+                return new ResponseEntity<>(exam, HttpStatus.OK);
+            }else{
+                return new ResponseEntity<>("Exam not found", HttpStatus.NOT_FOUND);
+            }
         }else{
-            return new ResponseEntity<>("Exam not found", HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
         }
+
 
   }
   @DeleteMapping("delete/{id}")
     public ResponseEntity<?> delete(@PathVariable int id) {
+      Authentication authentication= SecurityContextHolder.getContext().getAuthentication();
+      String email=authentication.getName();
+      if(getloginservice.getAuth(email)){
+          if(examservice.deleteExam(id)){
+              return new ResponseEntity<>(HttpStatus.OK);
+          }else{
+              return new ResponseEntity<>("Exam not found", HttpStatus.NOT_FOUND);
+          }
+      }else{
+          return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+      }
 
-        if(examservice.deleteExam(id)){
-            return new ResponseEntity<>(HttpStatus.OK);
-        }else{
-            return new ResponseEntity<>("Exam not found", HttpStatus.NOT_FOUND);
-        }
